# server/src/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  #  title: Mentor Success API
  description: API for the Mentor Success application
  contact:
    email: whitney@aiontechnology.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: "https://lj6z99n04d.execute-api.us-west-2.amazonaws.com/mentorsuccess-test-service-deployment"
...
paths:
...
  /api/v1/interests:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: header
        name: Accept
        schema:
          type:string

    # OPTIONS for /api/v1/interests
    options:
      tags:
        - resource
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/interests
    get:
      summary: Get interests
      description: Get all interests
      tags:
        - resource
      operationId: getInterests
      responses:
        200:
          description: Returned all interests
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestModelCollection'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/interests"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.Accept: "method.request.header.Accept"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

    # POST for /api/v1/interests
    post:
      summary: Add an interest
      description: Add an interest with the given data
      tags:
        - resource
      operationId: addInterest
      requestBody:
        description: The added interest
        content:
          application/json:
            schema:
              type: object
      responses:
        201:
          description: The new interest
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestResultModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/interests"
        responses:
          201:
            statusCode: "201"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "POST"
        type: "http"

  # /api/v1/interests/{interestId} -------------------------------------------------
  /api/v1/interests/{interestId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: header
        name: Accept
        schema:
          type:string
      - in: path
        name: interestId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired interest

    # OPTIONS for /api/v1/interests/{interestId}
    options:
      tags:
        - resource
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # PUT for /api/v1/interests/{interestId}
    put:
      summary: Update an interest
      description: Update an interest with the given data
      tags:
        - resource
      operationId: updateInterest
      requestBody:
        description: The updated interest
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: The new set of interests
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/interests/{interestId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.interestId: "method.request.path.interestId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "PUT"
        type: "http"

  # /api/v1/leadership_skills ------------------------------------------------------
...
components:
  responses:
    BadRequest:
      description: There was an error while creating the person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestModel'
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    Error:
      description: Encountered a server error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthenticated:
      description: Access token is missing or invalid
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthorized:
      description: User does not have permissions for operation
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotFound:
      description: Requested resource could not be found
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
...
  schemas:
...
    BadRequestModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        error:
          type: object
        message:
          type: string
        path:
          type: string

    # BehaviorModelCollection ------------------------------------------------------
...
    Empty:
      type: object

    # GameModel --------------------------------------------------------------------
...
    InterestModel:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 734aa3e8-52e4-488b-8922-09d8eb803ad8
        name:
          type: string
          maxLength: 50
          example: Cars

    # InterestResultModel ----------------------------------------------------------
    InterestResultModel:
      type: object
      allOf:
        - $ref: '#/components/schemas/InterestModel'
      properties:
        _links:
          type: object
          allOf:
            - $ref: '#/components/schemas/SelfLink'

    # InterestModelCollection ------------------------------------------------------
    InterestModelCollection:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            interestModelList:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/NameModel'

    # InvitationModel --------------------------------------------------------------
...
    NameModel:
      required:
        - name
      type: object
      properties:
        name:
          type: string

    # PersonModel ------------------------------------------------------------------
...
    SelfLink:
      type: object
      properties:
        self:
          allOf:
            - $ref: '#/components/schemas/UrlArray'

    # StudentModel -----------------------------------------------------------------
...
    UrlArray:
      type: array
      items:
        type: object
        properties:
          href:
            type: string
            format: url